RTOSDemo binary contains ARM Thumb instructions! No wonder so much of it is invalid.

+++ src/Arch/Avr/MemoryOperand.cs
@@ -1,6 +1,6 @@
 ﻿#region License
 /* 
- * Copyright (C) 1999-2017 John Källén.
+ * Copyright (C) 1999-2018 John Källén.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -28,8 +28,8 @@ namespace Reko.Arch.Avr
 {
     public class MemoryOperand : MachineOperand
     {
-        public RegisterStorage Base ;
-        public int Displacement;
+        public RegisterStorage Base;
+        public short Displacement;
         public bool PreDecrement;
         public bool PostIncrement;
