+++ src/Core/Operators/BinaryOperator.cs
@@ -1,6 +1,6 @@
 #region License
 /* 
- * Copyright (C) 1999-2017 John KÃ¤llÃ©n.
+ * Copyright (C) 1999-2018 John Källén.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -35,7 +35,7 @@ namespace Reko.Core.Operators
             return c1.IsValid && c2.IsValid;
         }
 
-        protected Constant BuildConstant(DataType t1, DataType t2, int val)
+        protected Constant BuildConstant(DataType t1, DataType t2, long val)
 		{
 			PrimitiveType p1 = (PrimitiveType) t1;
 			PrimitiveType p2 = (PrimitiveType) t2;
